+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 🆕 New features:
 
 🔧 Fixes:
+
+## 0.0.3
+
+💥 Breaking changes: workspaces.yaml is no longer valid
+
+🆕 New features: hbs template and json workspace data file will generate the yaml on the fly.
+
+🔧 Fixes: none.